updated Global Display to use itwin viewer
